Activity to model function
