+++ fedireads/activitypub/base_activity.py
@@ -1,16 +1,20 @@
 ''' basics for an activitypub serializer '''
 from dataclasses import dataclass, fields, MISSING
 from json import JSONEncoder
-from typing import Dict
+
+from django.db.models.fields.related_descriptors \
+        import ForwardManyToOneDescriptor
 
 
 class ActivityEncoder(JSONEncoder):
+    ''' allows conversion to JSON '''
     def default(self, o):
         return o.__dict__
 
 
 @dataclass
 class Image:
+    ''' image block '''
     mediaType: str
     url: str
     type: str = 'Image'
@@ -18,6 +22,7 @@ class Image:
 
 @dataclass
 class PublicKey:
+    ''' public key block '''
     id: str
     owner: str
     publicKeyPem: str
@@ -41,7 +46,38 @@ class ActivityObject:
             setattr(self, field.name, value)
 
 
+    def to_model(self, model):
+        ''' convert from an activity to a model '''
+        if not isinstance(self, model.activity_serializer):
+            raise TypeError('Wrong activity type for model')
+
+        model_fields = {}
+        for mapping in model.activity_mappings:
+            value = getattr(self, mapping.activity_key)
+            model_field = getattr(model, mapping.model_key)
+
+            # remote_id -> foreign key resolver
+            if isinstance(model_field, ForwardManyToOneDescriptor):
+                fk_model = model_field.field.related_model
+                value = resolve_foreign_key(fk_model, value)
+
+            model_fields[mapping.model_key] = value
+        return model.objects.create(**model_fields)
+
+
     def serialize(self):
+        ''' convert to dictionary with context attr '''
         data = self.__dict__
         data['@context'] = 'https://www.w3.org/ns/activitystreams'
         return data
+
+
+def resolve_foreign_key(model, remote_id):
+    ''' look up the remote_id on an activity json field '''
+    if hasattr(model.objects, 'select_subclasses'):
+        return model.objects.select_subclasses().filter(
+            remote_id=remote_id
+        ).first()
+    return model.objects.filter(
+        remote_id=remote_id
+    ).first()
